filesystem: document _pre_shutdown and link to systemd-shutdown impl
We decided to revert the use of systemd-shutdown for exporting zpools
because it didn't address the issue we were trying to solve. However,
dropping scripts for systemd-shutdown is an interesting implementation
from an architecture standpoint, because it can be executed even if
Subiquity does not cleanly exit.

Let's document the _pre_shutdown function and link to the original
systemd-shutdown implementation that we came up with for exporting
zpools earlier.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit f36c828debaea61e6649c3f938e92f57f6b94016)